refactor: move settings screen to feature module
