+++ lib/screens/home_screen.dart
@@ -4,7 +4,7 @@ import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import '../widgets/notes_tab.dart';
 import 'chat_screen.dart';
 import 'note_list_for_day_screen.dart';
-import 'settings_screen.dart';
+import '../features/settings/presentation/settings_screen.dart';
 import 'voice_to_note_screen.dart';
 import 'package:alarm_domain/alarm_domain.dart';
 import '../pandora_ui/palette_bottom_sheet.dart';
